requirements.txt
Django~=4.2.1
djangorestframework~=3.14.0
djoser~=2.2.0
pip~=23.1
wheel~=0.40.0
cryptography~=40.0.2
Pillow~=9.5.0
asgiref~=3.7.2
sqlparse~=0.4.4
psycopg2-binary~=2.9.6
colorama~=0.4.3
PyYAML~=5.4.1
requests~=2.26.0
certifi~=2023.5.7
six~=1.14.0
idna~=3.4
urllib3~=1.26.16
cloudinary~=1.33.0
setuptools~=67.8.0
future~=0.18.3
tzdata~=2023.3
phonenumbers~=8.13.11
python-dotenv~=1.0.0